+++ example/spec/parallel/ingestGranule/IngestGranuleSuccessSpec.js
@@ -13,7 +13,6 @@ const intersection = require('lodash/intersection');
 const isObject = require('lodash/isObject');
 
 const {
-  Execution,
   Granule,
   Pdr,
 } = require('@cumulus/api/models');
@@ -62,7 +61,9 @@ const {
   getTEARequestHeaders,
 } = require('@cumulus/integration-tests/api/distribution');
 const { LambdaStep } = require('@cumulus/integration-tests/sfnStep');
+const { getExecution } = require('@cumulus/api-client/executions');
 
+const { waitForApiStatus } = require('../../helpers/apiUtils');
 const {
   buildAndExecuteWorkflow,
   buildAndStartWorkflow,
@@ -117,7 +118,6 @@ describe('The S3 Ingest Granules workflow', () => {
   let beforeAllError;
   let collection;
   let config;
-  let executionModel;
   let expectedPayload;
   let expectedS3TagSet;
   let expectedSyncGranulePayload;
@@ -149,8 +149,6 @@ describe('The S3 Ingest Granules workflow', () => {
 
       process.env.GranulesTable = `${config.stackName}-GranulesTable`;
       granuleModel = new Granule();
-      process.env.ExecutionsTable = `${config.stackName}-ExecutionsTable`;
-      executionModel = new Execution();
       process.env.system_bucket = config.bucket;
       process.env.ProvidersTable = `${config.stackName}-ProvidersTable`;
       process.env.PdrsTable = `${config.stackName}-PdrsTable`;
@@ -318,11 +316,13 @@ describe('The S3 Ingest Granules workflow', () => {
 
   it('triggers a running execution record being added to DynamoDB', async () => {
     if (beforeAllError) throw SetupError;
-
-    const record = await waitForModelStatus(
-      executionModel,
-      { arn: workflowExecutionArn },
-      ['running', 'completed']
+    const record = await waitForApiStatus(
+      getExecution,
+      {
+        prefix: config.stackName,
+        arn: workflowExecutionArn,
+      },
+      ['running']
     );
     expect(['running', 'completed'].includes(record.status)).toBeTrue();
   });
@@ -745,9 +745,12 @@ describe('The S3 Ingest Granules workflow', () => {
 
     it('triggers the successful execution record being added to DynamoDB', async () => {
       if (beforeAllError || subTestSetupError) throw SetupError;
-      const record = await waitForModelStatus(
-        executionModel,
-        { arn: workflowExecutionArn },
+      const record = await waitForApiStatus(
+        getExecution,
+        {
+          prefix: config.stackName,
+          arn: workflowExecutionArn,
+        },
         'completed'
       );
       expect(record.status).toEqual('completed');
@@ -755,9 +758,12 @@ describe('The S3 Ingest Granules workflow', () => {
 
     it('triggers the failed execution record being added to DynamoDB', async () => {
       if (beforeAllError || subTestSetupError) throw SetupError;
-      const record = await waitForModelStatus(
-        executionModel,
-        { arn: failingWorkflowExecution.executionArn },
+      const record = await waitForApiStatus(
+        getExecution,
+        {
+          prefix: config.stackName,
+          arn: failingWorkflowExecution.executionArn,
+        },
         'failed'
       );
       expect(record.status).toEqual('failed');
@@ -835,7 +841,6 @@ describe('The S3 Ingest Granules workflow', () => {
       let cmrLink;
       let publishGranuleExecution;
       let updateCmrAccessConstraintsExecutionArn;
-
       let subTestSetupError;
 
       beforeAll(async () => {
@@ -1033,9 +1038,12 @@ describe('The S3 Ingest Granules workflow', () => {
           } else if (reingestBeforeAllError) {
             fail(reingestBeforeAllError);
           } else {
-            const reingestExecution = await waitForModelStatus(
-              executionModel,
-              { arn: reingestExecutionArn },
+            const reingestExecution = await waitForApiStatus(
+              getExecution,
+              {
+                prefix: config.stackName,
+                arn: reingestExecutionArn,
+              },
               'completed'
             );
             expect(reingestExecution.asyncOperationId).toEqual(asyncOperationId);