+++ packages/api/tests/endpoints/providers/update-provider.js
@@ -17,16 +17,23 @@ const {
   ProviderPgModel,
 } = require('@cumulus/db');
 const { Search } = require('@cumulus/es-client/search');
-const { bootstrapElasticSearch } = require('@cumulus/es-client/bootstrap');
+const {
+  createTestIndex,
+  cleanupTestIndex,
+} = require('@cumulus/es-client/testUtils');
 
 const models = require('../../../models');
 const {
   createFakeJwtAuthToken,
   fakeProviderFactory,
   setAuthorizedOAuthUsers,
+  createProviderTestRecords,
 } = require('../../../lib/testUtils');
 
 const assertions = require('../../../lib/assertions');
+const { put } = require('../../../endpoints/providers');
+const { buildFakeExpressResponse } = require('../utils');
+
 const testDbName = randomString(12);
 
 process.env.ProvidersTable = randomString();
@@ -44,9 +51,6 @@ const { app } = require('../../../app');
 const { migrationDir } = require('../../../../../lambdas/db-migration');
 
 let providerModel;
-const esIndex = randomString();
-let esClient;
-
 let accessTokenModel;
 let jwtAuthToken;
 
@@ -58,11 +62,17 @@ test.before(async (t) => {
 
   await s3().createBucket({ Bucket: process.env.system_bucket }).promise();
 
-  const esAlias = randomString();
-  process.env.ES_INDEX = esAlias;
-  await bootstrapElasticSearch('fakehost', esIndex, esAlias);
+  const { esIndex, esClient } = await createTestIndex();
+  t.context.esIndex = esIndex;
+  t.context.esClient = esClient;
+  t.context.esProviderClient = new Search(
+    {},
+    'provider',
+    t.context.esIndex
+  );
 
   providerModel = new models.Provider();
+  t.context.providerModel = providerModel;
   await providerModel.createTable();
 
   const username = randomString();
@@ -73,7 +83,6 @@ test.before(async (t) => {
   await accessTokenModel.createTable();
 
   jwtAuthToken = await createFakeJwtAuthToken({ accessTokenModel, username });
-  esClient = await Search.es('fakehost');
 });
 
 test.beforeEach(async (t) => {
@@ -92,7 +101,7 @@ test.after.always(async (t) => {
   await recursivelyDeleteS3Bucket(process.env.system_bucket);
   await accessTokenModel.deleteTable();
   await providerModel.deleteTable();
-  await esClient.indices.delete({ index: esIndex });
+  await cleanupTestIndex(t.context);
   await destroyLocalTestDb({
     knex: t.context.testKnex,
     knexAdmin: t.context.testKnexAdmin,
@@ -166,9 +175,21 @@ test('PUT updates existing provider', async (t) => {
       postgresOmitList
     )
   );
+
+  const updatedEsRecord = await t.context.esProviderClient.get(
+    testProvider.id
+  );
+  t.like(
+    updatedEsRecord,
+    {
+      ...expectedProvider,
+      updatedAt: actualProvider.updatedAt,
+      timestamp: updatedEsRecord.timestamp,
+    }
+  );
 });
 
-test('PUT updates existing provider in Dynamo and PG with correct timestamps', async (t) => {
+test('PUT updates existing provider in all data stores with correct timestamps', async (t) => {
   const { testProvider, testProvider: { id } } = t.context;
   const expectedProvider = omit(testProvider,
     ['globalConnectionLimit', 'protocol', 'cmKeyId']);
@@ -191,6 +212,9 @@ test('PUT updates existing provider in Dynamo and PG with correct timestamps', a
     t.context.testKnex,
     { name: id }
   );
+  const updatedEsRecord = await t.context.esProviderClient.get(
+    testProvider.id
+  );
 
   t.true(actualProvider.updatedAt > updatedProvider.updatedAt);
   // createdAt timestamp from original record should have been preserved
@@ -198,6 +222,8 @@ test('PUT updates existing provider in Dynamo and PG with correct timestamps', a
   // PG and Dynamo records have the same timestamps
   t.is(actualPostgresProvider.created_at.getTime(), actualProvider.createdAt);
   t.is(actualPostgresProvider.updated_at.getTime(), actualProvider.updatedAt);
+  t.is(actualPostgresProvider.created_at.getTime(), updatedEsRecord.createdAt);
+  t.is(actualPostgresProvider.updated_at.getTime(), updatedEsRecord.updatedAt);
 });
 
 test('PUT returns 404 for non-existent provider', async (t) => {
@@ -258,4 +284,187 @@ test('PUT without an Authorization header returns an Authorization Missing respo
     id: t.context.testProvider.id,
   });
   t.is(provider.globalConnectionLimit, t.context.testProvider.globalConnectionLimit);
+});
+
+test('put() does not write to PostgreSQL/Elasticsearch if writing to Dynamo fails', async (t) => {
+  const { testKnex } = t.context;
+  const {
+    originalProvider,
+    originalPgRecord,
+    originalEsRecord,
+  } = await createProviderTestRecords(
+    t.context,
+    {
+      host: 'first-host',
+    }
+  );
+
+  const fakeProvidersModel = {
+    get: () => Promise.resolve(originalProvider),
+    create: () => {
+      throw new Error('something bad');
+    },
+  };
+
+  const updatedProvider = {
+    ...originalProvider,
+    host: 'second-host',
+  };
+
+  const expressRequest = {
+    params: {
+      id: updatedProvider.id,
+    },
+    body: updatedProvider,
+    testContext: {
+      knex: testKnex,
+      providerModel: fakeProvidersModel,
+    },
+  };
+
+  const response = buildFakeExpressResponse();
+
+  await t.throwsAsync(
+    put(expressRequest, response),
+    { message: 'something bad' }
+  );
+
+  t.deepEqual(
+    await providerModel.get({
+      id: updatedProvider.id,
+    }),
+    originalProvider
+  );
+  t.deepEqual(
+    await t.context.providerPgModel.get(t.context.testKnex, {
+      name: updatedProvider.id,
+    }),
+    originalPgRecord
+  );
+  t.deepEqual(
+    await t.context.esProviderClient.get(
+      originalProvider.id
+    ),
+    originalEsRecord
+  );
+});
+
+test('put() does not write to Dynamo/Elasticsearch if writing to PostgreSQL fails', async (t) => {
+  const { testKnex } = t.context;
+  const {
+    originalProvider,
+    originalPgRecord,
+    originalEsRecord,
+  } = await createProviderTestRecords(
+    t.context,
+    {
+      host: 'first-host',
+    }
+  );
+
+  const fakeproviderPgModel = {
+    upsert: () => Promise.reject(new Error('something bad')),
+  };
+
+  const updatedProvider = {
+    ...originalProvider,
+    host: 'second-host',
+  };
+
+  const expressRequest = {
+    params: {
+      id: updatedProvider.id,
+    },
+    body: updatedProvider,
+    testContext: {
+      knex: testKnex,
+      providerPgModel: fakeproviderPgModel,
+    },
+  };
+
+  const response = buildFakeExpressResponse();
+
+  await t.throwsAsync(
+    put(expressRequest, response),
+    { message: 'something bad' }
+  );
+
+  t.deepEqual(
+    await providerModel.get({
+      id: updatedProvider.id,
+    }),
+    originalProvider
+  );
+  t.deepEqual(
+    await t.context.providerPgModel.get(t.context.testKnex, {
+      name: updatedProvider.id,
+    }),
+    originalPgRecord
+  );
+  t.deepEqual(
+    await t.context.esProviderClient.get(
+      originalProvider.id
+    ),
+    originalEsRecord
+  );
+});
+
+test('put() does not write to Dynamo/PostgreSQL if writing to Elasticsearch fails', async (t) => {
+  const { testKnex } = t.context;
+  const {
+    originalProvider,
+    originalPgRecord,
+    originalEsRecord,
+  } = await createProviderTestRecords(
+    t.context,
+    {
+      host: 'first-host',
+    }
+  );
+
+  const fakeEsClient = {
+    index: () => Promise.reject(new Error('something bad')),
+  };
+
+  const updatedProvider = {
+    ...originalProvider,
+    host: 'second-host',
+  };
+
+  const expressRequest = {
+    params: {
+      id: updatedProvider.id,
+    },
+    body: updatedProvider,
+    testContext: {
+      knex: testKnex,
+      esClient: fakeEsClient,
+    },
+  };
+
+  const response = buildFakeExpressResponse();
+
+  await t.throwsAsync(
+    put(expressRequest, response),
+    { message: 'something bad' }
+  );
+
+  t.deepEqual(
+    await providerModel.get({
+      id: updatedProvider.id,
+    }),
+    originalProvider
+  );
+  t.deepEqual(
+    await t.context.providerPgModel.get(t.context.testKnex, {
+      name: updatedProvider.id,
+    }),
+    originalPgRecord
+  );
+  t.deepEqual(
+    await t.context.esProviderClient.get(
+      originalProvider.id
+    ),
+    originalEsRecord
+  );
 });