+++ packages/db/src/index.ts
@@ -3,6 +3,7 @@ export {
   createTestDatabase,
   deleteTestDatabase,
   destroyLocalTestDb,
+  fakeAsyncOperationRecordFactory,
   fakeCollectionRecordFactory,
   fakeExecutionRecordFactory,
   fakeFileRecordFactory,
@@ -62,10 +63,15 @@ export {
 export { translateApiCollectionToPostgresCollection } from './translate/collections';
 export {
   translateApiProviderToPostgresProvider,
+  translatePostgresProviderToApiProvider,
 } from './translate/providers';
 export { translateApiRuleToPostgresRule } from './translate/rules';
-export { translateApiExecutionToPostgresExecution } from './translate/executions';
+export {
+  translateApiExecutionToPostgresExecution,
+  translatePostgresExecutionToApiExecution,
+} from './translate/executions';
 export { translateApiGranuleToPostgresGranule } from './translate/granules';
+export { translateApiPdrToPostgresPdr } from './translate/pdr';
 
 export {
   upsertGranuleWithExecutionJoinRecord,